requirements.txt
Flask==2.2.2
werkzeug==2.2.2
line-bot-sdk
# Pillow
googletrans==3.1.0a0
bs4
imgbbpy
zhdate
zhconv
pandas
